move tape properties and methods to quantumscript class
