+++ pennylane/__init__.py
@@ -34,6 +34,7 @@ import pennylane.qnn
 import pennylane.resource
 import pennylane.templates
 import pennylane.qchem
+import pennylane.qscript
 from pennylane.qchem import taper, symmetry_generators, paulix_ops, taper_operation, import_operator
 from pennylane._device import Device, DeviceError
 from pennylane._grad import grad, jacobian
