+++ tests/tape/test_tape.py
@@ -2197,7 +2197,7 @@ class TestOutputShape:
             qml.probs(wires=[0])
             qml.probs(wires=[1, 2])
 
-        with pytest.raises(TapeError, match="multiple probability measurements"):
+        with pytest.raises(ValueError, match="multiple probability measurements"):
             tape.shape(dev)
 
     def test_raises_multiple_different_measurements(self):
@@ -2212,7 +2212,7 @@ class TestOutputShape:
             qml.sample(qml.PauliZ(0))
 
         with pytest.raises(
-            TapeError,
+            ValueError,
             match="contains multiple types of measurements is unsupported",
         ):
             tape.shape(dev)
@@ -2228,7 +2228,7 @@ class TestOutputShape:
             qml.state()
             qml.density_matrix(wires=0)
 
-        with pytest.raises(TapeError, match="multiple state measurements is not supported"):
+        with pytest.raises(ValueError, match="multiple state measurements is not supported"):
             tape.shape(dev)
 
     def test_raises_sample_shot_vector(self):
@@ -2242,7 +2242,7 @@ class TestOutputShape:
             qml.sample()
 
         with pytest.raises(
-            TapeError,
+            ValueError,
             match="returning samples along with a device with a shot vector",
         ):
             tape.shape(dev)
@@ -2400,7 +2400,7 @@ class TestNumericType:
             qml.probs(wires=[0])
 
         with pytest.raises(
-            TapeError,
+            ValueError,
             match="Getting the numeric type of a tape that contains multiple types of measurements is unsupported.",
         ):
             tape.numeric_type
